qdoc: Delay ignoring of qt_xxx names until output
The ignoreSymbol() function in class ClangVisitor was told
to return true for every name that starts with "qt_". This
was too aggressive, so the test is now done at output time.
If a function name that starts with "qt_" has no documentation,
qdoc does not warn with "No documentation for qt_..."
because almost all such names are not to be documented. But
they can be documented, so that's why the test is performed
at the last moment.

Change-Id: I2f322bed32dcae43336ead0735490560190d8095
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>