qdoc: Hide clang errors for \fn commands in prepare phase
When clang parsing errors were detected in the prepare phase, they were
being printed without the preceding qdoc warning that shows where they
occur in the source file. The qdoc warning was not printed because qdoc
was running in the prepare phase, but the clang errors were printed, and
that was wrong. This update ensures that the clang parsing errors are only
printed in the generate phase (or in singleexec mode).

Change-Id: I011fa06626a5e9930fde329113d2bbac8830473b
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>