qdoc: Fix regression for wrongly reporting multiple topics
We have the case where multiple different topic commands can
appear in a qdoc comment in a .cpp file, when C++ is being
documented as QML. Specifically, a \qmlpropertygroup command
can be used with multiple \qmlproperty commands. The clang
code parser was reporting this as an error, even though clang
would never see these commaqndes because they are parsed by
the old qdoc command parser. The hasTooManyTopics() function
was changed to allow this case. It fixes many qdoc warnings
in QtDeclarative.

Change-Id: Iaea9a4b9c531f022544749819f65a48111ad0cc1
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>