qdoc: Generate correct links to QML properties
When a QML type is marked \qmlabstract, the documentation for
its properties is included on the reference page for each QML
type that inherits the abstract QML type. This means that when
using the link command to link to one of these properties in
HTML, qdoc must first obtain the correct file name. It was not
getting the correct file name, when the link command appeared
in documentation for a QML type that didn't inherit the QML
type marked abstract. This update corrects that bug.

Change-Id: I91ccb34e43618b7bbc1909026cbdf0c1b719df87
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>