qdoc: Don't output undocumented functions from Q_OBJECT
Earlier, a test was added to qdoc to ensure that functions added
to class declarations by macros like Q_OBJECT would not cause qdoc
to output warnings that the functions were not documented. But the
undocumented functions were still listed in the documentation. This
update keeps them out of the documentation.

Task-number: QTBUG-66742
Change-Id: I10014341fc7b870ef937ef7f0a303ccc301bf8b5
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>