qdoc: Canonicalize include paths
Resolve canonical include paths from the ones passed to QDoc from the
command line and .qdocconf variable 'includepaths'.

This allows adding include paths that are relative to the documentation
project (.qdocconf file):

  includepaths = -I .

which is useful for modules where we do not have a complete set of
include paths provided by qmake.

Change-Id: Ieaf0816e8c26857c873e4b521164c70bb1f5e5b3
Reviewed-by: Martin Smith <martin.smith@qt.io>