Qt Designer/generate_ui tool: Observe licenses of source files
Replace the hardcoded licenses in the .xsl files and replace by a
special @LICENSE@ keyword.
Read out the license from the source files in qttools/qtbase
respectively and create temporary .xsl files with the keywords
replaced accordingly.

Use the same mechanism for the #include directive in the ui4.cpp source.

Change-Id: I8a7937013041397954a30c676a9b699e0332a1e0
Reviewed-by: Kai Koehne <kai.koehne@qt.io>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>