macdeployqt: Correctly handle bundling debug Qt libraries with @rpath
After rpath support was introduced, we lost the ability to explicitly
bundle the debug version of the Qt libraries. We now make sure to
always run install_name_tool so that the library dependencies are
updated to the correct version, even in the case of @rpath.

After 27239f4fcfa6f64d in qtbase, we also now directly link to the
debug versions of Qt libraries when building the application in
debug mode, so the code has been taught to handle this case, by
not assuming that the library suffix is determined only by the
-use-debug-libs command line argument.

Task-number: QTBUG-48800
Change-Id: Ia7e058fa4d041fa1a7b8bdedc594750ee1f4cbfd
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>