winutils: Add possibility to run elevated background processes
In order to add localhost server support for winrt applications, we need
to be able to run elevated applications in the background. As a
preparation for this next patch, the new function is introduced. As
winrtrunner is not supported on OSs other than Windows, it is fine to
omit the implementation of this function on these operating systems for
now.

Change-Id: Iedb74a391e07a2b637f13bef84ad50c746dd8f0a
Reviewed-by: Andre de la Rocha <andre.rocha@qt.io>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>