winrtrunner: Add possibility to enable localhost capability for servers
Enabling localhost access for servers needs a loopbackexempt debug
session to be running while the application runs. This debug session
has to be started with elevated rights so the user might be asked for
confirmation, before the application can be launched.

This feature was introduced with Windows 10.0.14393.0.

[ChangeLog][winrtrunner][feature] Added ability to enable loopback exemption
for servers. This basically enables UWP applications to react on socket
connections that are made from the same machine for debugging purposes.

Change-Id: Iddf0c88657607075d85c4d9b326c8cd8b9771fa8
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Andre de la Rocha <andre.rocha@qt.io>