winrtrunner: Use dynamic shared memory size
We cannot hardcode the shared memory's size if we want to support large
qdebug streams. So a new message type was introduced, that signals the
need for a larger message buffer. If this message typ is received, the
shared memory has to be closed, so that a new shared memory can be
created by qtmain.

Task-number: QTBUG-61911
Change-Id: I0a3ecb9da4c7d8c1502443615ee73f6a140ef1e3
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>
Reviewed-by: Andre de la Rocha <andre.rocha@qt.io>