Add support for translation function aliases that are member expressions
When traversing the AST for call expressions, it may happen that the
base of the call expression is a field member expression instead of a
plain identifier. For example with CustomTranslator.myTr() the base is a
field member, with an identifier as base and myTr as name. We can
construct a qualified name from that quite easily and this way support
qualified tr aliases in qml files.

Task-number: QTBUG-67278
Change-Id: I106243d8eed6d8cec75819a44942cf750de2829a
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>