linguist: Add Regular expression checkbox in Find Dialog
Now it is possible to search using a Perl-like regular
expression. This enables the user to perform powerful
searches. There are many scenarios in which this is
convenient, let me show an example:

Source (Eng)     | Translation (Spa)
------------------------------------
From               Desde
From               De
... from ...       ... de ...
... of ...         ... de ...
from 1 to 10       de 1 a 10
(plus many other strings also containing "de" or "desde")

Even simple Regular Expressions like
^Desde$
^De$
from \d+ to \d+
become very convenient for fine-grained searches.

If the regular expression written in the Find Dialog
is not valid, text is highlighted in red and Find
button is disabled.

Change-Id: I90bb55d190d46263ec6adc27256d9c62a69e7f87
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>