linguist: Add actions in View menu to show more/fewer guesses
Three actions have been added to the main View menu (Guesses submenu)
in order to easily change the number of Guesses displayed.

View Menu
......
Guesses > Show more (alt - +)        [shows 5 more guesses]
          Show fewer (alt - -)       [shows 5 fewer guesses]
          ---
          Reset to default (alt - 0) [resets to 5 guesses]
......

Motivation behind this commit is that in large UIs with thousands
of strings, 5 guesses are not enough in some situations to gather
all the information necessary to correctly translate a string.

Change-Id: I3fdc1b8fc27af40300897cdfd1f9fa3620caad55
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>