linguist: Show remaining global&context translation counts in tooltips
Linguist shows the number of finished units and the number of editable
units in the lower right corner of the window (e.g.: 6714/6833).
It also shows this completion progress number per context.

With this change, it displays the difference between these numbers
in a tooltip (i.e. hovering the mouse on the count 6714/6833), for
both global progress count and per context progress count.

While this change may be arguable, the remaining count is the number
that a translator tracks and is constantly checking, other numbers
are less relevant. This small change improves User Experience since
the translator doesn't need to use a calculator or do a mental
calculation, which takes time and may also lead to miscalculations.

Change-Id: Icbb5e73bb4430d28f07eeff790b186579c48c8fb
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>