qdoc: Query environment variables from qmake
If qgetenv() doesn't return a valid environment variable, try querying
from qmake (assuming the variable name starts with 'Q').

This is useful for projects outside Qt, as it allows using e.g.
QT_INSTALL_DOCS and QT_INSTALL_HEADERS. The former can be used for
accessing the Qt documentation template, and the latter is useful for
projects that need to pass Qt's include paths to Clang (which QDoc
uses for parsing C++ documentation). The build system doesn't
provide these paths when trying to build docs for a module that
doesn't build on the host platform; for example, building Active Qt
documentation on Linux.

Change-Id: I758a5067d3dfdf13cadf979476323faea519ee40
Reviewed-by: Martin Smith <martin.smith@qt.io>