winrtrunner: Handle testlib specific output parameter style
QTestLib has a special way of giving the output parameter. Giving
-o filename,xml should output the test result to filename using the xml
scheme. To be able to run auto test in CI, we need to handle this scheme
and also support absolute file names as output.

To support the scheme, we can just check for the comma and remove the
part after it from the local file path.

In order to support absolute file paths, the application temporarily
writes its output into the default location. winrtrunner then copies the
result to its final destination.

Change-Id: I275444e1c3323301a1afd02aa8991a9e3fb4396c
Reviewed-by: Miguel Costa <miguel.costa@qt.io>
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>
Reviewed-by: Andre de la Rocha <andre.rocha@qt.io>