Prospective fix for linking qdoc against libclang with MinGW
Pick the libclang installation from MinGW or MSVC specific directories
on Windows for static linkage, as they cannot be mixed.

And it turn out that we need the explicit static library linkage for
MinGW as well (or just not for MSVC, as I guess the .lib there is
self-contained).

Task-number: QTBUG-68064
Change-Id: Ie20c8910e0a2776940dcd78a1b28e36f5ce30947
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@qt.io>