Detect doc version number and fix jumping to the proper doc
Add version-name metadata pair for the new documentation generated by qdoc.
Use heuristics to detect version number when version inside
metadata is not specified (applies to old docs). We do that only
for qt documentation (we detect the qtproject prefix inside namespace).

In order to match the proper namespace we compare the version
numbers of the original location with the version number of the
candidate list and pick up the matching one. Otherwise we behave
like before.

Task-number: QTBUG-59363
Change-Id: Id9f3d7d91e53647e2ad8d485d96888e80ce4f5b2
Reviewed-by: Kai Koehne <kai.koehne@qt.io>