qdoc: Allow escaping of single words to prevent autolinking
This update allows the writer to prevent QDoc from automatically
making a word into a link. For example, because LayoutMirroring
looks like a name that should become a link to something on a type
reference page, QDoc finds that reference page and the appropriate
section and makes LayoutMirroring be a link to it. By prepending a
single '\' to it in the text like this: \LayoutMirroring QDoc does
not turn it into a link and just appends LayoutMirroring to the text.

Change-Id: Ied7653d1e8f9d56dd5d86d1c16e2aaedf0a88a42
Task-number: QTBUG-20537
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>