Introduce distancefieldgenerator
A tool that allows you to pick prepare .qdf files which are pregenerated
distance fields in binary form that can be loaded by Qt to improve
startup time.

We currently only support cmaps subtable formats 0, 4, 6, 10 and 12,
as these are either A. the most commonly used (4 and 12) or B. simple
to implement.

The tool writes a .ttf file which is a copy of the original file,
but with the addition of a "qtdf" table that follows sfnt conventions.
This way, the file will work as a normal font file, as well, which
means that glyphs that are not in the pregenerated cache can easily
be generated at runtime.

Task-number: QTBUG-69356
Change-Id: Ib99c2d62f65e65973a60da4b1aa632b7ed3b2794
Reviewed-by: Lars Knoll <lars.knoll@qt.io>