Make the "Copyright" field in attribution files optional
Currently, the "Copyright" field in copyright attribution files is
mandatory. However, this obligation should be lifted so we could later
use README.chromium files with the qtattributionscanner tool (these
files do not contain a Copyright field ; in these cases, the copyright
owners can be found in the license file's body).

Change-Id: I569f8cd5d3f20fc85e5fcb873601053327841dda
Reviewed-by: Kai Koehne <kai.koehne@qt.io>
Reviewed-by: hjk <hjk@qt.io>