QDoc: Check aggregates with Node::nodeNameLessThan to improve sorting
Certain corner cases were not covered by the current implementation,
causing the sort to fail. This change aims to make QDoc capable of
sorting all nodes derived from aggregate, because an aggregate can be
expected to have a user visible title that can be used to sort content
alphabetically. Impressively, this has been working for most lists of
classes in the generated docs, just not all.

Task-number: QTBUG-67530
Change-Id: I5b5b19011c6a1e561894f8155c971da60aff7e2f
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Martin Smith <martin.smith@qt.io>