Extract qmake parser from lupdate and lrelease
Move the qmake parser from lupdate/lrelease to the new tool lprodump
that extracts the project structure from qmake projects. Its output is a
JSON file with an array of Project objects, which are defined as
follows.

Project ::= {
    string projectFile     // Name of the project file.
    string codec           // Source code codec. Valid values are
                           // currently "utf-16" or "utf-8" (default).
    string[] translations  // List of .ts files of the project.
    string[] includePaths  // List of include paths.
    string[] sources       // List of source files.
    string[] excluded      // List of source files, which are
                           // excluded for translation.
    Project[] subProjects  // List of sub-projects.
}

We can pass such a project description to lupdate and lrelease via the
-project argument.

Also, introduce the tools lupdate-pro and lrelease-pro. Both tools are
replacements for the QMake mode of lupdate/lrelease and call lprodump
to generate a temporary project description which is then passed to
lupdate/lrelease.

Passing .pro/.pri files to lupdate/lrelease is deprecated as of now.

As a side-effect of this change, we enable lrelease-pro to traverse
SUBDIRS projects recursively.

Task-number: QTBUG-35997
Task-number: QTBUG-48416
Change-Id: I256e9d36169306d02817262c68cf1bc410ea82de
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>