qdoc: Fix generated C++ and QML type lists
As of change 03b8f9d9, for nodes read from an index, QDoc creates a Doc
instance that evaluates as empty (Node::hasDoc() returns false). This
resulted in 'All C++ Classes' and 'All QML Types' pages to have no
content as the code that selects the listed nodes had a check against
an empty Doc.

Remove this check as unnecessary: for a type to be listed it's enough
for it to be marked as public.

Task-number: QTBUG-69693
Change-Id: I35a61c99ef3d44023601705be26e74e695cd1d0e
Reviewed-by: Martin Smith <martin.smith@qt.io>