Fix linking of qdoc against libclang when cross-compiling to QNX/Android
When cross-compiling to QNX, qt_find_clang.prf is evaluated against the
target configuration, so the check for "mingw" to select the right
LLVM_INSTALL_DIR variant does not work as that's the host configuration.

So instead we must formulate the conditions differently and rely on the
fact that we are either using MSVC where the same compiler is used for
host and target architecture (even if the architectures differ) or
otherwise minwg is the host compiler (while target may be something
else).

Task-number: QTBUG-69793
Change-Id: I31acc127c94e0a8570242d1c8ea0db873f9fd253
Reviewed-by: Simo Fält <simo.falt@qt.io>