Fix linking libclang with mingw when cross-compiling
Commit 4ec82667f632992e827299c96d3bfd2b221a029b moved the

    mingw: CLANG_LIBS += -ldl

condition from a option(host_build) .pro file to qt_find_clang.prf which
is evaluated in target scope. That means when cross-compiling, we can't
use mingw anymore as a way to detect the host compiler.

Instead, rely on MinGW being the only alternative to MSVC that we
support as a host compiler when cross-compiling. The msvc part is dealt
with earlier in the file.

Also cleaned up the LLVM_INSTALL_DIR determination that suffers from the
same symptom (commit ac6de68fe7d4fef251c101ef25812fe2198925f8), where
the msvc bit needs a proper scope for the else to work correctly.

Task-number: QTBUG-69793
Change-Id: Id544c7929b79fc3c8763d50e0bac4a3a689ade60
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>