qdoc: Fix generation of namespace documentation
Commit a36f925d refactored the Section-related code, resulting
in different data structures used for storing sections for
C++ classes and namespaces. However, the HTML generator
still assumed that the same structure could be used for both,
resulting in invalid/no output for namespace documentation.

This commit fixes it by selecting the correct vector based
on whether we're processing a class or a namespace node.

Task-number: QTBUG-70321
Change-Id: Ia80dec9193e22c453350588aba10f4c033f6ba79
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Martin Smith <martin.smith@qt.io>