qdoc: Allow parameters for \code, \qml, \js, and \badcode commands
Allow optional parameters for commands that enclose code snippets.
This is intended for replacing simple strings within the snippet,
typically using a macro that extends to certain parameter(s).

Within the snippet, the parameter locations are marked similarly
to .qdocconf macro parameters:

\code foo bar
\1+\2=\1\2
\encode

Renders "foo+bar=foobar"

A code snippet parameter cannot contain spaces, as space is used
as a delimiter. If no parameters are passed, the snippet is
rendered as-is.

Task-number: QTBUG-67818
Change-Id: I500e6b1a9095c9c55f47c8d75951792f31745545
Reviewed-by: Martin Smith <martin.smith@qt.io>