QDoc: Move global variables out of main.cpp
This change moves most of the global variables out of main.cpp, and
into the new qdocglobals class, with getter and setter functions for
each variable.

It also introduces the basic structure for autotests for QDoc.

The motivation is to increase the testability of QDoc, as well as to
simplify any future refactoring.

Change-Id: If71ba9f1837e759aa75b1c9fec326d72c41c4fcf
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>
Reviewed-by: Ole-Morten Duesund <olemd@odinprosjekt.no>
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>