qdoc: Accept all #include lines in module header
qdoc was building a temporary module header by copying in the actual
module header and then following it with includes for all the private
headers in the module's private include directory. But this attempt
to include headers that end in "_p.h" didn't include any headers at
all, so that code is removed in this update.

The problem with just copying the entire module heaqder into the
temporary module header was that some of the #include lines that
qdoc needed to see were not parsed by clang because they were
surreounded by #if QT_CONFIG(feature) if the feature wasn't
present for the Qt build.

Instead of copying the module header to the tempory module header,
this update changes it to read the module header line by line and
write each #include to the temporary module header, including the
ones that are surrounded by QT_CONFIG(feature):

_#if QT_CONFIG(dtls)
_#include "qdtls.h"
_#endif

This reduces the qdoc error count, but the example above causes a
static assert when the dtls feature is not present. So this must
be added as well:

_#ifndef Q_CLANG_QDOC
 QT_REQUIRE_CONFIG(dtls);
_#endif

Change-Id: I7fe26b5b714d41654a2cfdda19b8fe4b31999b91
Reviewed-by: Martin Smith <martin.smith@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>