qdoc: Do not print error for future functions
Some function declarations are marked as not existing until Qt 6.0
in both the .h and .cpp files, but they are documented in the .cpp
file. qdoc can't tie the documentation for these functions to their
declarations because of the way clang is used, and qdoc reports
that reports the documentation as an error.

This update let's qdoc ignore these functions and not print the error
when they are marked with \since 6.0. This will also work with other
future versions.

This update also collects all the uses of #define COMMAND_xxx in one
location, which was partially required by the \since 6.0 change.

Change-Id: I55052359f387406da340c748768f8e76c0b39d53
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>