Assistant: Fix a crash when removing the docs
Inside the PreferencesDialog, when the currentRegisteredDocsSelection()
returns the unsorted list we may crash inside removeDocumentation().
Please note that we iterate through the currentSelection() list in the
reverse order. However, when the values are not sorted there, we may remove
items from entries list in a random order. After the first removal
the row index may point outside of the list or point to the wrong doc.

The fix is to sort the list returned by the currentRegisteredDocsSelection().
In this way we enure that we remove entries in the reverse order
and guarantee that values of the currentSelection list matche the indices
of the entries list.

Fixes: QTBUG-71399
Change-Id: I545f358e1d55e64823ffdaf0ce27a8b97aeedf04
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit 10263f42de0d2d1fb731deb4213b76fa6c0acd0f)