Qt Designer: Port to new configure system
Replace the old QT_NO... define checks by the configure features where
applicable.
Join the sections for the QTimeEdit/QDateEdit (inheriting
QDateTimeEdit).

For the printer, introduce a define depending on the availability
of the printsupport library and QPrintDialog (cf textedit example)
to accommodate platforms that don't build QtPrintSupport.

Task-number: PYSIDE-797
Change-Id: I752d99a19b55616fa2ac5c36fad803d6d861e327
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>