Qt Designer: Overhaul the layouts manual
In the introductory paragraph, add a link to the
Layout Management overwiew.

Distinguish between top level layouts and layout objects
and move the section on top level layout above the other.
Flesh out the sections on layouts and top level layouts.

Mention the Shift-Click selection and on that occasion fix the
shortcut for multiselection, which was fixed to conform to standards
in Qt 4.

Change-Id: Id8d8e28c77cbec07c7ac816e4425a8ce337b28c1
Fixes: QTBUG-70841
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>