qdoc: Fix link resolution for types under namespaces
When QDoc tries to resolve a link to a type from a function parameter,
it sets a 'TypesOnly' find flag. For types that contain a scope
resolution operator(s), i.e. double-colons, QDoc first looks for the
scope (namespace), and then the type under that scope.

The problem was that the TypesOnly flag was passed to also the
function call to find the namespace. This failed as namespaces are
not types.

Another problematic corner case happened with name clashes between
a C++ module and a namespace, as is the case with e.g. 'Qt3DCore'.
Here, QDoc tried to resolve the module node as the scope, and
subsequently failed to find anything under that scope as module nodes
have no children.

Task-number: QTBUG-69484
Change-Id: Ia3c8fa96263a24176a11cf23a3469a374de7973a
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Martin Smith <martin.smith@qt.io>