Qt Designer: Fix crash when reloading forms with resources
Change qtools/548a75049efaa3f9b3087016e3e36cc9b4f23e2e caused
icon/pixmap properties to leak when setting a null object on the
property editor since the code removing the properties was not
executed.

Default to the active form window when trying to find a form window
for the object in question and bail out with a warning if that fails.

Fixes: QTBUG-71809
Task-number: QTBUG-68507
Change-Id: I3fd96dd9b72d822c47071286b641254530f1568d
Reviewed-by: Kai Koehne <kai.koehne@qt.io>