Fix documentation for cacheDirectory
"collection" and "default" are alternatives. In Qt 4, the documentation
therefore used '|' for separation, the '>' character was - probably
by mistake - introduced in mass commit 7727e2075.

Anyhow, the effect of base="default" is the same as not setting the
attribute at all, so I guess this is what most people should do, and
we should also show as example.

Change-Id: Idb0ce083dedae7b2cda5f45eef2e5ab2f8ae3ab0
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>