winrtrunner: Pass logging rules set in the environment to the app
If an auto test fails Coin tries to rerun the same test with logging
enabled to get additional information. Instead of just enabling extended
logging for winrtrunner we have to pass these rules to the application
that is run by winrtrunner to be able to obtain the information.

As we cannot set the application's environment from winrtrunner we write
qtlogging.ini where the application expects it to be so that logging is
enabled on startup.

Change-Id: I238a083f6d499f5b6b5eadb43545a5857050a2e8
Reviewed-by: Andre de la Rocha <andre.rocha@qt.io>
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>