winrtrunner: Do not accept fully wildcarded logging rules from environment
When an auto test fails in CI, it does another run using
QT_LOGGING_RULES=*=true. For winrt builds this rule would enable extended
logging for winrtrunner and the output would be flooded with winrtrunner
specific information.

Instead of confusing the users, we disregard the logging rules in
winrtrunner if they do not specifically enable the application's logging
so that CI runs will only show information that might be interesting to
the user.

Change-Id: If51a666315a2e33b83d6a7613a65ab4a152eddc4
Reviewed-by: Andre de la Rocha <andre.rocha@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>