Add simple smoke test for QDoc
This adds a simple QDoc project that makes use of a subset of
QDoc's commands in a .qdoc file, and a data set containing the
expected output when qdoc runs on this project. The idea is to
generate output that can be compared across versions of QDoc
itself.

Task-number: QTBUG-71168
Change-Id: Idaba28d3fdfd5dbedfcd2ce4cc4d8e9361bd2008
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>